backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-socketio[asyncio_client]==5.10.0
aiohttp==3.9.1
deepgram-sdk==3.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
python-dotenv==1.0.0

# Zoom SDK alternatives - to be evaluated
# zoomus==1.1.5
